Streaming/StreamingApp/stream_requirements.txt
blinker==1.8.1
click==8.1.7
Flask==3.0.3
Flask-MySQL==1.5.2
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mysql-connector-python==8.1.0  # Prefer mysql-connector-python over mysql-connector
numpy==1.26.4
pandas==2.2.2
PyMySQL==1.1.0
pyspark==3.0.0  # Ensure compatibility with pyspark-stubs
pyspark-stubs==3.0.0.post3
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tzdata==2024.1
Werkzeug==3.0.2
sqlalchemy